Added XcodeSettings class
