+++ lib/branch_io_cli/configuration.rb
@@ -2,3 +2,4 @@ require "branch_io_cli/configuration/configuration"
 require "branch_io_cli/configuration/report_configuration"
 require "branch_io_cli/configuration/setup_configuration"
 require "branch_io_cli/configuration/validate_configuration"
+require "branch_io_cli/configuration/xcode_settings"
